requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
pytesseract==0.3.10
Pillow==10.1.0
PyPDF2==3.0.1
python-magic==0.4.27
flask==3.0.0
python-dotenv==1.0.0
langdetect==1.0.9
regex==2023.10.3